Push entity history test env. to sub-package level
